file 101-quote.c content change

+++ 0x00-hello_world/101-quote.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	char message = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
-	unsigned int length = sizeof(message);
+	size_t length = sizeof(message);
 	write(2, message, length);
 	return (1);
 }
